add argument for user and error handling if argument or database doesn't exists

+++ scripts/after_deploy.sh
@@ -1,9 +1,13 @@
 #!/bin/bash
-
-cd ~/missed_call_mail_alert/calllist
+if [ ! $1 ] ; then
+  USER="pi"
+fi
+cd /home/$1/missed_call_mail_alert/
 
 # backup your database
-cp calllist.db calllist.db.old
+if [ -f calllist.db ]; then
+   cp calllist.db calllist.db.old
+fi
 
 # make sone scripts executable
 chmod +x missed_call_mail_alert.sh
@@ -11,3 +15,4 @@ chmod +x webfrontend/server.py
 chmod +x analyse.py
 chmod +x database.py
 chmod +x pingscript.sh
+date >> deploy.txt
